moss: Move indexing code out of cli folder & provide pub interface
Such that indexing can be called from other projects
